ECV-552 Split Section1MaterialsExporter
