+++ src/EPR.Calculator.Service.Function/Startup.cs
@@ -243,8 +243,9 @@ internal static class ServiceRegistration
         services.AddTransient<ICalcResultParameterOtherCostExporter, CalcResultParameterOtherCostExporter>();
         services.AddTransient<ICalcResultModulationExporter, CalcResultModulationExporter>();
         services.AddTransient<ICalcResultCommsCostExporter, CalcResultCommsCostExporter>();
-        services.AddTransient<IProducerIdentityExporter, ProducerIdentityExporter>();
-        services.AddTransient<IProducerSummaryExporter, ProducerSummaryExporter>();
+        services.AddTransient<ICalcResultSummaryPartExporter, ProducerIdentityExporter>();
+        services.AddTransient<ICalcResultSummaryPartExporter, Section1MaterialsExporter>();
+        services.AddTransient<ICalcResultSummaryPartExporter, ProducerSummaryExporter>();
         services.AddTransient<ICalcResultSummaryExporter, CalcResultSummaryExporter>();
         services.AddTransient<IBillingFileJsonWriter, BillingFileJsonWriter>();
         services.AddTransient<ICalcResultLateReportingExporter, CalcResultLateReportingExporter>();
